ルートを統合し、レストラン、ルーム、タベログのAPIエンドポイントを新しいファイルに分割
